requirements.txt
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
numpy==1.21.2
matplotlib==3.5.1
Pillow==9.0.1
tqdm==4.62.3
argparse==1.4.0
